# src/future-be/src/swagger.json
{
	"components": {
...
		"schemas": {
			"AddressRes": {
				"properties": {
					"province": {
						"type": "string"
					},
					"district": {
						"type": "string"
					},
					"ward": {
						"type": "string"
					},
					"specificAddress": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"receiver": {
						"type": "string"
					},
					"_id": {
						"type": "string"
					},
					"default": {
						"type": "boolean"
					}
				},
				"required": [
					"province",
					"district",
					"ward",
					"specificAddress",
					"phone",
					"receiver",
					"_id",
					"default"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateAddressDTO": {
				"properties": {
					"province": {
						"type": "string"
					},
					"district": {
						"type": "string"
					},
					"ward": {
						"type": "string"
					},
					"specificAddress": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"receiver": {
						"type": "string"
					}
				},
				"required": [
					"province",
					"district",
					"ward",
					"specificAddress",
					"phone",
					"receiver"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_CreateAddressDTO_": {
				"properties": {
					"province": {
						"type": "string"
					},
					"district": {
						"type": "string"
					},
					"ward": {
						"type": "string"
					},
					"specificAddress": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"receiver": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UpdateAddressDTO": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Partial_CreateAddressDTO_"
					},
					{
						"properties": {
							"default": {
								"type": "boolean"
							}
						},
						"type": "object"
					}
				]
			},
...
			"Pick_ICreateUser.name_": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UserResDTO": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Pick_ICreateUser.name_"
					},
					{
						"properties": {
							"_id": {
								"type": "string"
							}
						},
						"required": [
							"_id"
						],
						"type": "object"
					}
				]
			},
			"ICreateUser": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"username",
					"password",
					"avatar"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ISignJWT": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"role": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserInfo": {
				"properties": {
					"name": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"birthday": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					}
				},
				"required": [
					"name",
					"avatar",
					"email",
					"birthday"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WishlistItem": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"thumbnail": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"isStock": {
						"type": "boolean"
					}
				},
				"required": [
					"_id",
					"thumbnail",
					"name",
					"price",
					"isStock"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CartItemResDTO": {
				"properties": {
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"thumbnail": {
						"type": "string"
					},
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"quantity",
					"price",
					"name",
					"thumbnail",
					"_id"
				],
				"type": "object"
			},
			"HttpStatus": {
				"enum": [
					200,
					201,
					202,
					400,
					401,
					403,
					404,
					406,
					408,
					500,
					501,
					502,
					503
				],
				"type": "number"
			},
			"AddToCart": {
				"properties": {
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"productId": {
						"type": "string"
					}
				},
				"required": [
					"quantity",
					"productId"
				],
				"type": "object"
			},
			"UpdateQuantity": {
				"properties": {
					"action": {
						"type": "string"
					}
				},
				"required": [
					"action"
				],
				"type": "object"
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "future-be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/addresses": {
			"post": {
				"operationId": "CreateAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/AddressRes"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Addresses"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAddressDTO"
							}
						}
					}
				}
			}
		},
		"/addresses/{id}": {
			"put": {
				"operationId": "UpdateAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/AddressRes"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Addresses"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAddressDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "string"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Addresses"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/authenticate/register": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/UserResDTO"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"login"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateUser"
							}
						}
					}
				}
			}
		},
...
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/UserResDTO"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateUser"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/IUserInfo"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": []
			}
		},
		"/users/sign-jwt-token": {
			"post": {
				"operationId": "SignJwtToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISignJWT"
							}
						}
					}
				}
			}
		},
		"/users/addresses": {
			"get": {
				"operationId": "GetMyAddresses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"items": {
														"$ref": "#/components/schemas/AddressRes"
													},
													"type": "array"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": []
			}
		},
		"/users/count-user": {
			"get": {
				"operationId": "CountUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "number",
													"format": "double"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": []
			}
		},
		"/users/wishlist": {
			"get": {
				"operationId": "GetWishlist",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"items": {
														"$ref": "#/components/schemas/WishlistItem"
													},
													"type": "array"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": []
			}
		},
		"/users/cart": {
			"get": {
				"operationId": "GetCart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"$ref": "#/components/schemas/CartItemResDTO"
											},
											"type": "array"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "AddToCart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/CartItemResDTO"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddToCart"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAllCart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "string"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": []
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "FindUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/UserResDTO"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/wishlist/{productId}": {
			"post": {
				"operationId": "InsertWishlistItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/WishlistItem"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteWishlistItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "string"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/HttpStatus"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/cart/{productId}": {
			"delete": {
				"operationId": "DeleteCart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "string"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateQuantity",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"properties": {
														"action": {
															"type": "string"
														},
														"productId": {
															"type": "string"
														}
													},
													"required": [
														"action",
														"productId"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateQuantity"
							}
						}
					}
				}
			}
		},
		"/users/setting": {
			"put": {
				"operationId": "UpdateInfo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/IUserInfo"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"birthday": {
										"type": "string"
									},
									"avatar": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}